+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/base/database/tables/PostsTable.kt
@@ -4,6 +4,7 @@ import com.github.insanusmokrassar.AutoPostTelegramBot.base.database.exceptions.
 import com.github.insanusmokrassar.AutoPostTelegramBot.base.database.exceptions.NoRowFoundException
 import com.github.insanusmokrassar.AutoPostTelegramBot.mediumBroadcastCapacity
 import com.github.insanusmokrassar.AutoPostTelegramBot.utils.NewDefaultCoroutineScope
+import com.github.insanusmokrassar.AutoPostTelegramBot.utils.chooseCapacity
 import com.github.insanusmokrassar.TelegramBotAPI.types.MessageIdentifier
 import kotlinx.coroutines.channels.BroadcastChannel
 import kotlinx.coroutines.launch
@@ -17,9 +18,9 @@ typealias PostIdMessageId = Pair<Int, MessageIdentifier>
 val PostsTableScope = NewDefaultCoroutineScope()
 
 object PostsTable : Table() {
-    val postAllocatedChannel = BroadcastChannel<Int>(mediumBroadcastCapacity)
-    val postRemovedChannel = BroadcastChannel<Int>(mediumBroadcastCapacity)
-    val postMessageRegisteredChannel = BroadcastChannel<PostIdMessageId>(mediumBroadcastCapacity)
+    val postAllocatedChannel = BroadcastChannel<Int>(chooseCapacity(mediumBroadcastCapacity))
+    val postRemovedChannel = BroadcastChannel<Int>(chooseCapacity(mediumBroadcastCapacity))
+    val postMessageRegisteredChannel = BroadcastChannel<PostIdMessageId>(chooseCapacity(mediumBroadcastCapacity))
 
     private val id = integer("id").primaryKey().autoIncrement()
     private val postRegisteredMessageId = long("postRegistered").nullable()