+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/plugins/rating/database/PostsLikesTable.kt
@@ -5,6 +5,7 @@ import com.github.insanusmokrassar.AutoPostTelegramBot.base.plugins.commonLogger
 import com.github.insanusmokrassar.AutoPostTelegramBot.mediumBroadcastCapacity
 import com.github.insanusmokrassar.AutoPostTelegramBot.smallBroadcastCapacity
 import com.github.insanusmokrassar.AutoPostTelegramBot.utils.NewDefaultCoroutineScope
+import com.github.insanusmokrassar.AutoPostTelegramBot.utils.chooseCapacity
 import com.github.insanusmokrassar.AutoPostTelegramBot.utils.extensions.subscribe
 import com.github.insanusmokrassar.TelegramBotAPI.types.ChatId
 import kotlinx.coroutines.channels.BroadcastChannel
@@ -22,9 +23,9 @@ private const val resultColumnName = "result"
 private val PostsLikesTableScope = NewDefaultCoroutineScope()
 
 class PostsLikesTable : Table() {
-    val likesChannel = BroadcastChannel<PostIdUserId>(smallBroadcastCapacity)
-    val dislikesChannel = BroadcastChannel<PostIdUserId>(smallBroadcastCapacity)
-    val ratingsChannel = BroadcastChannel<PostIdRatingPair>(mediumBroadcastCapacity)
+    val likesChannel = BroadcastChannel<PostIdUserId>(chooseCapacity(smallBroadcastCapacity))
+    val dislikesChannel = BroadcastChannel<PostIdUserId>(chooseCapacity(smallBroadcastCapacity))
+    val ratingsChannel = BroadcastChannel<PostIdRatingPair>(chooseCapacity(mediumBroadcastCapacity))
 
     private val userId = long("userId").primaryKey()
     private val postId = integer("postId").primaryKey()