+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/plugins/scheduler/PostsSchedulesTable.kt
@@ -6,6 +6,7 @@ import com.github.insanusmokrassar.AutoPostTelegramBot.extraSmallBroadcastCapaci
 import com.github.insanusmokrassar.AutoPostTelegramBot.plugins.base.PostsUsedTable
 import com.github.insanusmokrassar.AutoPostTelegramBot.smallBroadcastCapacity
 import com.github.insanusmokrassar.AutoPostTelegramBot.utils.NewDefaultCoroutineScope
+import com.github.insanusmokrassar.AutoPostTelegramBot.utils.chooseCapacity
 import com.github.insanusmokrassar.AutoPostTelegramBot.utils.extensions.subscribe
 import kotlinx.coroutines.channels.BroadcastChannel
 import kotlinx.coroutines.channels.ReceiveChannel
@@ -22,9 +23,9 @@ class PostsSchedulesTable : Table() {
     private val postId = integer("postId").primaryKey()
     private val postTime = datetime("postTime")
 
-    val postTimeRegisteredChannel = BroadcastChannel<PostIdPostTime>(smallBroadcastCapacity)
-    val postTimeChangedChannel = BroadcastChannel<PostIdPostTime>(extraSmallBroadcastCapacity)
-    val postTimeRemovedChannel = BroadcastChannel<Int>(extraSmallBroadcastCapacity)
+    val postTimeRegisteredChannel = BroadcastChannel<PostIdPostTime>(chooseCapacity(smallBroadcastCapacity))
+    val postTimeChangedChannel = BroadcastChannel<PostIdPostTime>(chooseCapacity(extraSmallBroadcastCapacity))
+    val postTimeRemovedChannel = BroadcastChannel<Int>(chooseCapacity(extraSmallBroadcastCapacity))
 
     init {
         PostsTable.postRemovedChannel.subscribe {